SDL_render_gles2: remove ineffective widening cast warning: either cast from 'int' to 'size_t' (aka 'unsigned long') is ineffective, or there is loss of precision before the conversion [bugprone-misplaced-widening-cast]

+++ src/render/opengles2/SDL_render_gles2.c
@@ -1812,7 +1812,7 @@ GLES2_RenderReadPixels(SDL_Renderer * renderer, const SDL_Rect * rect,
     int status;
 
     temp_pitch = rect->w * SDL_BYTESPERPIXEL(temp_format);
-    buflen = (size_t) (rect->h * temp_pitch);
+    buflen = rect->h * temp_pitch;
     if (buflen == 0) {
         return 0;  /* nothing to do. */
     }
